pyproject.toml
[project]
name = "pdf-study-notes"
version = "1.0.0"
description = "Page-by-page PDF note extraction for open-book exam prep"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.25.0",
    "httpx>=0.28.0",
    "rich>=13.9.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.12.0",
]

[project.scripts]
pdf-notes = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
